Public spec: Add todo about keeping possibly multiple states
I need to think this through. Maybe, due to the asynchronicity of method
calls and responses it is actually ok to model the state completely
linear from the public point of view (which would be great!). But I am
not sure if this is true.

Originally https://github.com/dfinity-lab/dfinity/commit/a0353228e3c26f7433a4ff7c7593ed2cbff6e760