Public spec: callbacks are required (dfinity-lab/dfinity#503)
the current phrasing of the abstract behaviour is broken, it may try to
invoke an unspecified callback. This fixes this, and brings it in line
with the prose further up.

It does not address the question of whether such callbacks _should_
be optional, and what that should mean.
Originally https://github.com/dfinity-lab/dfinity/commit/253328f7ba48d022500263db1d49d097e10f8889