Public Spec: Argument builder-sytle interface for `msg.reply` (dfinity-lab/dfinity#568)
* Public Spec: Argument builder-sytle interface for `msg.reply`

in Multi-function message send (dfinity-lab/dfinity#370) we introduced the notion of
`call.append_data` to allow the canister to copy out the data to be sent
incrementally. We obviously want to do the same thing for replies, for
symmetry reasons.

This is also a prerequisite for returning payments with replies
(see dfinity-lab/dfinity#561)

There is some duplication going on here, where for every kind of
argument we need a function to add it to a call and one to add it to a
reply. We could avoid that by having a “virtual reply callref”. We could
even have a separate `call.create_function` to create it. But this is
not completely satisfying, as it would not be creating a call (so it is
a misnomer) and it would not support the `call.on_reply` and
`call.on_reject` functions (which means that it is not correct
abstraction).

* Update docs/spec/public/index.adoc

Co-Authored-By: John Wiegley <johnw@newartisans.com>

Originally https://github.com/dfinity-lab/dfinity/commit/fba172a80b29e6ceba303a1e9a02db815cdd0ab7