Public Spec: TODO: Reads may address specific state (dfinity-lab/dfinity#576)
When we had this discussion a few weeks ago in Zurich, the conclusion was that we
will not provide synchronisation help for the user when doing read
queries, and leave it to the application to solve the problem if they
worry about that.

But in today's call with Mack, Stanely, Johan, John, I learned that we
_do_ want queries to be able to specify a state (likely a
very-near-future-state that the node will then just wait for to be
reached before responding).

There was text in the public spec that is not in agreement with that
requirement, so I rephrased that, and added a todo.
Originally https://github.com/dfinity-lab/dfinity/commit/25e768e78632c9cdaa97eff6dec6622f6b8d0505