Public Spec: JSON-RPC’ify the interface DFN-142 (dfinity-lab/dfinity#528)
this is based on input from dfinity-lab/dfinity#277 and more discussion (here and via video call)
with Stanley. It introduces the concept of (API) messages, which comes as asynchronous
ones (need to be submitted and their status checked separately) and synchronous
ones (which are essentially read requests, and do not modify state).

The RPC layer thus only has two end points, `submit` and `read`.

This also allowed me to simplify the presentation of the abstract behavior of
the system.

While introducing this, it was natural to also encode the two-step protocol
for creating a canister (i.e. register first to get a canister id, then install code
with canister).

We do _not_ lift canister methods to the level of API messages; instead, there
is a generic `call` API message.
Originally https://github.com/dfinity-lab/dfinity/commit/66ae05adefe7ede1542df56cc1b306484e9b1f71