Make “query call” an inherent property of a canister’s method (dfinity-lab/dfinity#614)
previously, I was phrasing query calls as something transparent to the
canister: It was up to the caller to choose to execute a call as a query
call if the caller is only interested in the result, but does not want
the state to be committed, essentially as a more convenient alternative
to running a shadow copy off-chain.

But Andreas told me that this is the wrong model, and canister should
_deliberately_ export methods as either update or query calls. So this
specs this out.

The main difference is that while a user can still call a query call
using the same path as update calls (e.g. going through consensus etc,
maybe in order for the user to get a certification of the result, or
simply as part of inter-canister messaging), if the canister declares
the method to be a query call, the state would not be committed. Which
is good and might be surprising otherwise.

We use the name of the export to distinguish update from query calls on the
System API level. One might expect that we could use the IDL here, but
so far the IDL is _not_ something that the system is concerned with, and
this is a decoupling worth preserving.

The spec rules out having the same method name be both an update and a
query method.

This commit also makes the abstract `CanisterModule` more precise, e.g. by
giving it types that by construction rule out state updates from query
calls.
Originally https://github.com/dfinity-lab/dfinity/commit/83dfef96e3f722364d4fb0897c5156084e876e6a