Public Spec: TODO about System API call→api message (dfinity-lab/dfinity#647)
During my visit to Zurich, during a meeting with Dom, I noticed that the
System API as envisioned is too limited: While the Public API supports
many kinds of API messages (update calls, but also a bunch of other
things like canister installation, eventually payment, account
queriying, admin stuff like setting permissions), the System API has a
provision only for canister calls.

But derived from the requirement that we eventually want to support
governance canisters, and also the (implicit) requirement that it should
in general be possible take any kind of outside (RPC API) interaction
and automate it using a canister, this is too limiting.

One could add new custom system calls for all the other kind of API
messages (aka requests) supported, but this does not scale and would
prevent canisters doing generic messaging.

I will therefore change the `callref` stuff to instead provide a way for
the canister to submit an (API) message, and get the response.

This is actually not critical for the DevNet, as inter-canister calls
are out of scope, but I wanted to note this down here.

(this was discussed with @rossberg in Zurich)
Originally https://github.com/dfinity-lab/dfinity/commit/78a97c069196b73d761f363e996147c1661f31f0