Public spec: Refine rejects, document reject codes (dfinity-lab/dfinity#644)
this is follow up from a meeting on Aug 8 with @rossberg, @allendfinity,
@akhi3030 and

It says that rejects carry both a reject code and a message, and that
`msg.reject` does _not_ indicate a code.

The “trap via system call” comes separately.

The list of reject codes is just an initial stab, I am happy to get
refinements here, and we can also keep refining that list on `master`.

I am using the term “reject code” instead of “error class” to avoid the
overloaded term “error” in general, and because it is a bit misleading
to talk about a “class” code when we do not expose the (system-internal)
codes that are _class_ified by that.
Originally https://github.com/dfinity-lab/dfinity/commit/00f264ce27b4fe58599a2307e44ff967fe16e102