Public spec: Request signatures (dfinity-lab/dfinity#746)
as per discussion with, in particular, @manudrijvers or @dubovitskaya,
we learn that we will Ed25519 for authentication of extenal users. And
although the preicse plan for how identities and keys are going to be
associated is still in the air, it makes sense to specify how request
signatures will look like, so that we can start implementing this.
Originally https://github.com/dfinity-lab/dfinity/commit/26e8480143def248f42243f2db6950390c2ab112