Public Spec: Make canister installation argument required (dfinity-lab/dfinity#1089)
@nmattia argues that it is inconsistent that canister calls have
mandatory arguments, but canister installation requests don’t, and while
the analogy is not perfect (a canister does not have to export
`dfn_init`), there is a good argument to treat them the same on the
platform level.

It would be on the SDK to know when a canister does not have a
`dfn_init`, or one that does not care about the argument, and then pass
an empty blob in the `arg` field without bothering the user.

The benefit is simpler types and less special casing inside the system.
Originally https://github.com/dfinity-lab/dfinity/commit/d41ea8e506cf7f3b5d62787ab8dea984305dce59