Public Spec: “Requests first” refactoring (dfinity-lab/dfinity#1175)
This is a refactoring of the Public Spec document structure: Instead of
mixing the list of requests, their fields etc together with rather
low-level details of the actual HTTP-based protocol, this separates the
two into

 * a section “Requests” introducing the notion of requests and the set
   of requests (including message ids), separate from

 * the section HTTPS Interface, that (generically) describes how
   requests are encoded, signed and submitted.

The goal of this, besides reducing the level of section nesting for most
of the document, is to pave the way towards a refactoring of the _System
API_ to work on the level of requests, as discussed in
https://dfinity.slack.com/archives/CN4SJRNG4/p1568270101016000
(Not done here yet, will be a separate PR eventually.)

This structure also makes it clearer that public key signatures play a
role only for _authentication_, but do not affect that actual meaning of
requests.  Therefore, they are not mentioned in the “Requests” section,
but only in the HTTPS Interface. This is the current direction of Mack
and Dom’s work on Schema; not yet set in stone, and obviously we can
refactor this document if their direction changes again.

The only actual user-visible change is that `sender_pubkey` no longer
contributes to the calculation of the `request_id`.
Originally https://github.com/dfinity-lab/dfinity/commit/dcd361b3a81bad9390060ba7ba73b9b68d12f24c