Public spec: Remove entropy API (dfinity-lab/dfinity#1273)
I learned from @jancamenisch that the entropy design is up in the air
again, and it is unlikely that the system will provide synchronous
access to randomness, as it would slow down _all_ canisters, and an
async interface might be likely to come. But in that case, the API
sketched here will _not_ work. So to avoid confusion, this PR removes
it.
Originally https://github.com/dfinity-lab/dfinity/commit/255a74df9d43c001a1afb334112c56cf2b01415f