Public Spec: Clarify request status evolution (dfinity-lab/dfinity#1486)
* Public Spec: Clarify request status evolution

this is based on a discussion with, I think, @jensgroth, @jancamenish
and others in Zurich two or three weeks ago. The main result is that we
have a way to characterize the `Accepted` and `Processing` request
states in terms of externally meaningful guarantees (and not in terms of
internal concepts).

@nmattia nudged me to finally write this up.

* s/Accepted/Received

* Fix adoc warnings

* A user creates a request. These days.

* Pointless but harmless

* Remove an edge from diagram

Originally https://github.com/dfinity-lab/dfinity/commit/4cf73ead65b8407edf0735a0549d945b9869dc20