Public Spec: Caller identification (dfinity-lab/dfinity#1551)
* Public Spec: Caller identification

it seem that waiting for the full spec around identities (and other
mostly unrealted things that are co-designed with that) should not stop
us from spec’ing and implementing the simple feature of
“A canister knows who called it.”

This adds the `ic.msg_caller` call to do that. No surprises here.

In this PR I also add an `sender` field to query calls, so that there is
something for `ic.msg_caller` to return in that case. See the note there
about (the lack of) authentication.

* Section “Authentication” applies to synchronous requests only

* Sign reads

lunch discussion with @jancamenisch indicates that we do care about an
attacker who wants to read other user’s data, but has not compromised a
node.

This implies that we need to also sign request status reads and canister
query calls, in order to authenticate the author.

* Apply suggestions from code review

Co-Authored-By: John Wiegley <john@dfinity.org>

* Implementation status

Originally https://github.com/dfinity-lab/dfinity/commit/c380b78c9cc9a041465d7e61d810d09a4f92c749