Public Spec: Change process to “Johan approves” (dfinity-lab/dfinity#1931)
* Public Spec: Change process to Johan approves

It wasn’t working very well that I was creating this document mostly
detached from the concerns of the Execution Team, and detached from the
overall release management and feature decisions of the whole
organization. Parts of the spec were out of sync with the implementation
for extended periods of time.

Johan @granstrom has the laudable goal to have the client catch up with
any spec have to be approved by Johan, changes within a week or two or
so. But this can only work if all spec changes correspond to something
that is required by the next release.

As Johan has much better insight into the overall release planning, he
will now be the gate keeper to the public spec. He will be the one that
merges changes to the public spec when and if they are actually about to
be implemented.

Features that are not agreed on can still be spec’ed out, but have to
live on a branch (e.g. a Draft PR) until they are decided to go in.

Editorial changes (i.e. those that do not change the meaning) should be
clearly marked as such, so that Johan knows he can approve them
independent of project management.

* Update docs/spec/public/index.adoc

Co-Authored-By: Jens Groth <46536510+JensGroth@users.noreply.github.com>

Originally https://github.com/dfinity-lab/dfinity/commit/463d2b88db0de28337524f29b5472b3c0b87f611