Public Spec: Merge sections “requests“ and “HTTPs interfaces” (editorial) (dfinity-lab/dfinity#1937)
* Public Spec: Merge sections “requests“ and “HTTPs interfaces” (editorial)

In dcd361b3a81bad9390060ba7ba73b9b68d12f24c / dfinity-lab/dfinity#1175 I moved the abstract
notion of requests into a separate section from the HTTPS interface, at
that time anticipating that they’d be used on the System API as well.

Shortly after the idea of the Virtual IC canister was born, which made
that change obsolete and not helpful. I didn't revert it right away,
planning to do that after the Virtual IC canister lands (dfinity-lab/dfinity#1438), which I
expected to land shortly after.

Unfortunately, the Virtual IC canister was stuck waiting for Dom.
But this left the Public Spec with aless-than-ideal structure for 7
weeks, and was causing actual confusion to users of this document.

This week I had to accept that the Virtal IC Canister design is shleved
for the forseeable future, so in order to avoid such confusion (and in
order to leave the document in a reasonable shape as I yield ownership
of it), this PR reverts dfinity-lab/dfinity#1175 (and fixes some unrelated asciidoc typos).

* fix link

* Apply suggestions from code review

Thanks to @jwiegley for proof-reading!

Co-Authored-By: John Wiegley <john@dfinity.org>

Originally https://github.com/dfinity-lab/dfinity/commit/f7f5710b34d9d5d809d3ba4aeb39c8fb2b8fc130