Revert "Public spec: Storing and retrieving assets (dfinity-lab/dfinity#1487)" (dfinity-lab/dfinity#2106)
This reverts commit 7ae6ffebf10a859f426be1b1bdcbba6ce638df11.

It seems that there was a hicup in our (still new) process and
@granstrom didn’t actually sign off dfinity-lab/dfinity#1487:
https://github.com/dfinity-lab/dfinity/pull/1487#issuecomment-565061730

This reverts the change. @granstrom, is that your intent?

Going forward, I’ll let the pigs on the table (SDK and Client) hash out
the high-level approach to assets, and will join the discussion when it
gets concrete enough to spell out the actual interface.
Originally https://github.com/dfinity-lab/dfinity/commit/39e8652650992be7657df1ac99bdadeb959661be