ic-stub: Remove ability to create canisters (#1154)
* ic-stub: Remove ability to create canisters

this was a “backdoor” to support code that was removed in #1140. In
order to keep the delta betwen ic-stub and the Public Spec small, in
both directions, this is removing the code. It was a hack anyway (e.g.
it was synchronous), the real deal will likely look different.

* Do not import stub.* system api

Originally https://github.com/dfinity-lab/motoko/commit/69f248b11b4229f84fa5a5dcd92f7e83c2929f0e