Public Spec: Refine user request structure (dfinity-lab/dfinity#2085)
Ingress messages are conceptually an "envelope" containing the content of the message, a signature on the (hash) of this content and a public key which can be used to verify the signature, reference slides. This hints to a distinction between the content of the message (which contributes to how the request id is calculated) and metadata of the request, such as the signature and the sender's public key, which are used for authentication of requests but do not contribute to the request id.

This change makes this distinction clearer, by separting the content of a request from the “signature envelope”. This encoding represents more naturally what a request comprises of and it leads to cleaner description of how the request id can be calculated and what are the metadata used for authentication.

Since the change affects the CBOR encoding of requests, it affects both the SDK and the Client.
Originally https://github.com/dfinity-lab/dfinity/commit/2cdd787ea3e1dc0d8cd417c4569d661016fcd9e0