nix/haskell-packages: replace IFD with importing generated files (#1220)
* nix/haskell-packages: replace IFD with importing generated files

Replace IFD calls to callCabal2nix with importing generated
`default.nix` files.

The generated `default.nix` files are stored in the repository and
contain instructions on how to regenerate them.

The generation of the `default.nix` files is done using the
`haskellSrc2nixWithDoc` function which generates a directory
containing a `default.nix` which is the result of running `cabal2nix`
with the `extraCabal2nixOptions` on the provided `src`. A header is
then added to the `default.nix` which contains instructions on how to
regenerate that file.

Finally the new jobs are added which check that the stored
`default.nix` matches the expected `default.nix`. So whenever the
`.cabal` files are updated and don't match the stored `default.nix`
anymore CI will complain.

Originally https://github.com/dfinity-lab/motoko/commit/b90fbc518837eb98fc3cc2a6fe59804ed429a9f6