Refactor setup for generated nix files (#1223)
the reiterates on #1220. The motivation for the refactoring was that the
generated files were not using `gitSource`, so I would get cache misses
and dirty files in the repo. Fixing this was easier if all generated nix
files are in one place, so I did the following changes:

 * All generated nix files are in `nix/generated`. This de-pollutes the
   normal working area with generated files, and keeps them out from
   sources (more precise derivations).
 * The local generated files now use `gitSource`.
 * The generator is now `nix/generate.nix`.
 * I moved complex logic from `nix/default.nix` to `nix/generate.nix`,
   this means everything is in one place, instead of spread over several
   nix files.
 * There is now only a single check to check if the generated files are
   upto date, and a single command to update the files.
 * This check is no longer marked `allowSubstitutes = false`. The check
    isn’t cheap (it requires `cabal2nix`) and we want the check to happen
    on builders, not hydra itself, so that the cache has `cabal2nix`.
 * The check is now included in `all-systems-go`.
Originally https://github.com/dfinity-lab/motoko/commit/290534b598883a6dae33e10bdf4618129ab5bb0c