Testing: Universal canister (#9)
Testing becomes much more expressive and frictionless when the test script can seamlessly define and send requests and specify the behaviour of the “other side” (the canister).

So let’s define a singe wasm module that takes instructions in the message argument what it should do, so that we can script it via arguments.

The canister is written in rust (and the nix setup around that is more complicated that I wish it would be – why can’t nixpkgs just not ship the wasm32-unknown-unknown target) in a separate directory/derivation. It implements a very simple stack-based langauge with binary opcodes.

The Haskell module `Ic.Test.Universal` provides a simple DSL to create these scripts, using
GADTs to ensure type safety of these scripts. No ASTs, or parsers, or compilers around, 
this is very simple (but works quite nicely).