Public Spec: The controller field (#14)
there was already some ground work done in
https://github.com/dfinity-lab/dfinity/pull/584
when this was still called an admin. This adds the necessary requests to
change the controller, as per the “What is a canister” document
(https://docs.google.com/document/d/1GMqwIOBi2Rp8wwOhJOrhVgHeSL2_bUt96_vJNV0lK30/edit#)

(The “What is a canister” document also envisions batching admin
commands in a single atomic request; that is somewhat orthogonal, so not
done in this PR.)

This adds a _single_ controller, not a set of controlllers; Doms
explicit vision is to use proxy canisters if you need multiple
controller.