Public Spec: The cleanup callback (#2)
The problem this solves is the following:

 * Inter-canister calls are asynchronous.
 * Canisters need to keep certain resources alive between the call and
   the callback.
 * They rely on the “a callback will be called” guarantee to eventually
   free these resources. This is a main feature of our platform.
 * We roll back state on traps.
 * Callbacks are likely to trap in certain situations, not all of which
   are under control by the canister (out of gas, bad IDL values)

This together leads to situations where the callbacks traps and the
resource is never freed.

One way to solve that is what is proposed here: If a reply or reject
callback traps for whatever reason, the canister gets a “second shot” in
the “cleanup callback”. In this callback, the canister can _only_ clean
up its own state (no replying/rejecting/sending calls/inspecting
arguments), and will be given a (yet to be determined) fixed gas limit
to do so.

For example in Motoko it would remove a closure from the list of GC
roots, that's approximately like reading 5 words and writing two. (Other
implementations may do something more complicated, so don’t set the gas
limit too close to that).

Access to `ic0.stable_` is allowed to match the eventual multi-memory
future where this access no longer goes through the System API.

The cleanup callback is specified the same way as the reply or reject
callbacks, i.e. a table entry+env pair. One could also imagine a fixed
`canister_cleanup` export that receives the `reply_env` and `reject_env`
arguments, but the same could be said about the reply and reject
continuations themselves, so this is more consistent.

Originally https://github.com/dfinity-lab/dfinity/commit/ee19f2a2b94a9bb16fd3a0a978c3ee72793f2ad8