ic-ref-test: Test reply after trap
The “Bidirectional Messaging” design means that even though a canister
traps in one callback, it may still respond (reply or reject) in another
one. As surfaced in https://github.com/dfinity-lab/dfinity/pull/3374 the
current replica does properly implement this, and I must admit that the
`ic-ref-test` test suite is incomplete in that respect; but no more.