Public Spec: Request signatures to include a domain separator (#20)
It is good practice to make sure that the signatures included in the
request are guaranteed to be made with the intention of signing such a
request (and not an email attachment or something else). Thefore, the
sign and verify functions should sign not just the request id, but also
a so-called domain separator.

This is based on a spec by @fspreiss and @andreacerulli.