Public Spec (formal section): GC call contexts (#74)
previously there was no need to clean up call contexts, but canister
stopping (#70) introduces one. To make that change easier to specify,
this adds a transition that removes call contexts that are no longer
needed.

It has similar conditions as the starvation rule. Which was buggy (due
to incomplete refactorings), so fixing that along the way.