Public Spec: Make all canister ids system-chosen (#75)
The idea that the registering party (user or canister) can determine the
id of the generated canister is not reconcilable with the present system
design, as it requires too much central authority. Problems mentioned by
@bogwar are:

1. we need some centralized store to ensure uniqueness
2. related to 1, we have a smaller design space for how to generate
   identifiers (and all benefits we may get by tailoring them in certain
   way, e.g. defining identifiers as subnetid+ctr)
3. we may need to store identifiers in perpetuity (if we don't want to
   allow any kind of reuse)
4. reusing identifiers from the development phase in the production
   phase is bound to create problems (source @ielashi)

If we can’t support this feature for the next release (and its
usefulness is in question in general), let’s take it out of the spec,
(and then cut a reduced release).

This still declares that the system should not _use_ these sender-derived
ids, so that we can add them later without any clahes.

This has buy in from @dfinity-bjoern and Jan, according to Bogdan.