Public Spec: Clarify that debug_print does not necessarily print (#78)
It already said that the data may need to be escaped (if binary), and
that it may not be printed to `stdout` but rather stored or logged. Now
it clarifies that the data is typicall not `\n` or `\0` terminated.
Maybe thiw would have avoided the confusion in the Rust libraries that led to
https://github.com/dfinity-lab/dfinity/pull/3936