Public Spec: Fix ic00 Candid syntax (#80)
when writing the Candid signature, I somehow assumed that Candid has the

    record { foo : int; nat }  =  record { foo : int; nat : nat }

short-hand. I am not sure where this is coming from, but Candid does
_not_ have this.

Unfortunately, the left hand side _is_ valid syntax, and means

    record { foo : int; nat }  =  record { foo : int; n : nat }

where `n = hash("foo") + 1`. The point of that shorthand is to encode
tuples i.e.

    record { int; nat; blob } = record { 0 : int, 1 : nat, 2 : blob }

but mixed with named fields it is, eh, very confusing.

This fixes the candid file to the intended meaning.