Public Spec: Expiry of requests (#68)
much of the explanation in section “Asynchronous requests” cannot really
be understood without the `expiry` field, and without the expiry field we
have worries about replay attacks. So high time to get this in.

The implementation will reject requests with a expiry too soon in the future
(else it will not reach `processing` in time) or too far (else it makes the system
keep stuff alive for too long). The precise bounds are, for now, implementation
defined.

Cherry-picked from 6861389b2c00034302b2db815336c237be2df583