Public spec: Emphasize diff between reject and trap (#170)
This PR aims at clarifying some aspect of the spec with respect to persistence and message rejection. It is not a "real" spec change in the sense that it does not introduce anything new.

## Description of the pre-existing ambiguity
Section https://docs.dfinity.systems/public/#_message_execution read, before this PR:

> If message execution succeeds, the state is update[d] and possible outbound calls and responses are enqueued.
>
> If message execution traps, or if it double-responds, the message gets dropped. No response is generated (as some other message may still fulfill this calling context).

It is not obvious in which of those categories the case of "execution did not trap; did not call reply; called ic0.reject exactly once" would fall here. The condition of the second paragraph does not include this case, so this paragraph provides no information about this case.

The first paragraph uses the verb "succeeds". This verb is not defined in this section. In https://docs.dfinity.systems/public/#_responding we can read:
> Eventually, the canister will want to respond to the original call, either by replying (indicating success) or rejecting (signalling an error):

This would suggest that "success" is defined by (a) not trapping (b) not calling `ic0.reject`, and (c) calling `ic0.reply` exactly once. However, this is __not__ the intended meaning in the section about message execution, where I know from talking to spec editors @nomeata that the case "not trapping; not calling `ic0.reply`; calling `ic0.reject` exactly once" __is__ considered a success from the point of view of message execution.

This PR aims therefore at disambiguating what "success" really means in either context.

## Resolution of the ambiguity
A survey of the occurrences of "succe" in the spec shows that, overwhelmingly, this word is associated with a message execution that called `ic0.reply` exactly once. Therefore the proposed resolution is to remove the use of the word "succe[ss]" in the sentence "If message execution succeeds, the state is update[d]...".

In section https://docs.dfinity.systems/public/#_nomenclature we can read:
> WebAssembly functions are exported by the WebAssembly module or provided by the System API. These are invoked and can either trap or return, possibly with a return value. Functions, too, have parameters and take arguments.

Therefore, the verb _returning_ should be used instead of _succeeding_.