Delegation from WebAuthn key to plain key (#175)
As browsers do not implement _silent_ authentication (yet), namely authentication in which the user does not need to interact with the token/sensor, using WebAuthn to authenticate all queries would require the user to interact with its device repeatedly. This PR introduces _delegation_, in which WebAuthn is merely used to sign an ephemeral key with a limited validity period.