Public Spec: Add ic-deposit_funds (#160)
This PR introduces a new IC method called `deposit_funds` which is meant to be used to top up a canister's balance, especially in cases where the canister is out of cycles and thus unable to process a normal message containing funds. The controller can always use this method to top up the canister's balance, whereas other canisters may use it only when the canister is "Frozen" (out of cycles).