Public Spec: Require `sender` field (#203)
In anonymous requests, the `sender` field was allowed to be either unset or set to `0x04`. This has lead to confusion, and complicates parsing the requests a bit. As there does not seem to be a strong reason to keep the field as optional, this PR makes the field required, and mandates that anonymous requests set `sender = 0x04`.

(cherry picked from commit 268420a426497c416e87b69796c487aa1643193c)