Add domain separator for web authentication (#215)
The previous version of the spec was ambiguous with respect to the use of domain separators in web authentication signatures. This PR clarifies that domain separators shall be used also for web authentication.

The PR also takes the description of signatures out of the authentication section and puts it into a separate section.