Public Spec: Certification with delegation (#218)
this PR adds the necessary bits to allow agents to validate a
certificate produced by a non-root subnet.

This completes #90 and #143 (which so far only work securely on the root
subnet). It was extracted from the Topology PR (#178), but is more
important than that, and it’s good to take smaller steps.