ic-ref: Fix ECDSA
because of a bug in `ic-ref-test`, this was not actually tested. This
hid a bug in `ic-ref`, and also hid the fact that we were using a
different encoding for the signatures than the replica does.

So with this change, instead of DER-encoding the public key (as common
in bitcoin), we use a raw encoding of `r` followed by `s`, in big-endian
format.