Public Spec: No default memory allocation (#251)
as per call with @akhi3030 and @bogwar:

The current default of 8GB causes huge headache, as canister storage is
scarce (especially since in the worst case, subnets have to store a
multitudes of state copies). With a default of 8GB, subnets would fill
up too quickly even if the canister admins never intended to let the
canisters grow.

In the eventual design, these allocations are reservations, i.e. you are
guaranteed to be able to grow up to your reservation, any further grows
is satisfied on-demand if possible. In that design, the natural default
is `0` (no guarantees, but also no worries until subnets actually get
full – great for small canisters, canisters that eventually have
constnat memory use).

Unfortunately, that is _not_ the current semantics: Currently, the
system will prevent memory grows beyond the allocation, and fixing that
(although not huge) will not happen before ☿₁. Nevertheless, the 8GB
default is too bad so that the IC wants to drastically reduce the
default allocation (to 10MB, I heared), and switch to the reservation
semantics soon afterwards. This decision, I was assured, was made in
considerations of the DX degregations (canisters suddenly failing to
allocate memory; possibly even in pre-upgade) and the possible bad rep
effects.

This change simply stops pretending that the Public Spec sets the
default, and instead makes it an implementation-defined costant.