ic-ref: Serialize the state (#263)
This uses [`serialise`](https://hackage.haskell.org/package/serialise) to serialize the state of the IC to a CBOR-encoded file. It’s CBOR, but not the interoperable kind, so it is not intended to be parsed by other tools.

The design of `IC.Ref` leads to rather bloated files: 
 * All requests are persisted, even when they are old
   Can be alleviated by exposing the “real” time to `IC.Ref` and implementing request expiry.
 * Modules are duplicated (once for the canister, and once in the request)

But simply compressig/decompressing on the fly helps a lot here.

The logic about when the state is written is a bit involved, but nicely contained in `src/IC/StateFile.hs`.

When it can’t read the file (corrupted, or different version) `ic-ref` fails to start, and tells the user what file to delete.