Public Spec: Add support for secp256k1 (#264)
This PR adds support for ECDSA on `secp256k1` for ingress messages.

One new requirement is that Neurons given to seed round investors shall be controlled by the keys that the seed round investors used when registering with the Ethereum smart contract. Those keys are on the `secp256k1` curve and used in ECDSA.

Note that, beyond prescribing ECDSA, the signatures actually differ from Ethereum signatures. Ethereum uses SHA-3 and encodes the signatures as a triplet `r, s, v` instead of as a pair `r, s`.