Editorial: Prepare public spec for publishing (#265)
this commit gets the public spec into publishable form. Changes are:

 * The Preamble, which was addressing internal developers, was moved
   to the repository’s `README.md`
 * Removed my name
 * Improved introduction, to clarify target audience and set
   expectations.
 * Removed a few notes that speculate about the future
 * A little bit of copy-editing.
 * Checked that the document does not contain any internal links

I do not think we should do larger rewrites at this point: This would
cause pain for us as we juggle the various branchs and in-flight
changes, and I don’t think the public needs a different document than
what we have used internally so far. Yes, it has its flaws, and we
should fix them eventually, but I would not tie that to making it
public.