ic-ref-test: Reinstall test checks that state is reset
I learned the hard way (i.e. in “production”) that the replica, upon
`reinstall`, is not resetting the memory properly. Unfortunately, I did
not test for this here… now we do. It does _not_ trigger the bug in the
replica (probably the Rust-implemented canister initializes its state in
`(start)`), but the extra checks also do not hurt.