Spec: Replace rate limiting by ingress filtering (#285)
This PR reflects recent changes in the replica implementation: the ingress rate limiting mechanism was removed and replaced by a simpler ingress inspection mechanism, in which the canister can only decide to accept or reject a message. The mechanism was implemented in [recent PR](https://github.com/dfinity-lab/dfinity/pull/8496).

The rate limiting logic we implemented previously was rather involved. It also required the replicas to keep many past states of all canisters, as otherwise the notaries were not able to check that the pre-consensus execution was correct. We decided to replace it for Beta with a simpler mechanism for basic DoS protection. Rate limiting remains a goal for the IC and is likely to return in a more resource-friendly form.