ic-ref(-test): Implement and test canister_inspect_message (#287)
This implements the design in #285.

Because there is only one universal canister, it always exposes the
`canister_inspect_message` method. The default action is to always
accept, of course.

This change makes the test suite more liberal with regard to errors on
submit: For some tests it now expects either a HTTP error (4xx or 5xx)
on submission, or a rejection later on. Maybe at some point we can make
the spec more precise here.

The API availability code is extended to check both ingress and
inter-canister calls for `U` and `Q`.