Github action to create antora tree on separate branch
This adds a script that transforms the interface specification file to
the right directory structure that is expected by antora, the tool that
we use to fill the sdk website.

It also adds a github action (see .github/workflows/antora.yml) that pushes
the result of this to branch `public-antora`.