Spec: deposit_cycles(): any caller allowed (#296)
This PR removes the restriction that only the controller can deposit cycles on a canister.

The discussion around the requirement was (re)opened in the context of the cycles minting canister: https://github.com/dfinity-lab/dfinity/pull/9193#discussion_r590208955

Johan clarified that the requirement that canisters explicitly accept incoming cycles applies to update requests but not necessarily for management calls: for the latter, the intent of the caller is obvious so no need to be cautious.

(cherry picked from commit 2373a38e9742a7ec3c97c7f5d83f5c664a14b231)