Spec: HTTP v2 for stateless rounting (minimal version) (#294)
This proposal describes the interface changes needed to allow stateless routing on edge nodes (#248), thus paving the way to full access decentralization (#246)

This does **not** change the the replicated parts of ingress message processing. The `call` request is untouched, so validation, authentication, scheduling, status tracking etc. work as before.

On the replica-side this only affects the HTTP Handler.

The changes are:

 * The HTTP endpoints now include the effective canister id in the path  (`/api/v2/canister/<canisterid>/call`, where `canisterid` is the textual representation), even for fetching the status of a request. The _effective_ canister id is the callee for normal calls, but the acted-upon-canister for management calls.

   This means that an endge node  (i.e. what is currently done by `ic-fe`) no longer has to store a mapping from request id to canisters, which means it can be stateless, which means there can be many of them; crucial for decentralization.

   It also means that `ic-fe` no longer has to look at the request body, addressing an important concern by Nik (#242).

 * The categorization in `submit` and `read` requests doesn’t make sense if we only have three requests in total. This simplifies the representation and has `/call`, `/read_state` and `/query` again. This causes lots of mostly editorial changes.

The introduction of the “effective canister id” may seem a bit hacky, but the logic behind it is relatively small and can (and should) be encapsulated in the agents, so not something that developers have to worry much about. The HTTP interface is inherently more low level than the canister interface, and thus it is not surprising that agents have to do some extra work to raise the level of abstraction for their users.

This proposal is a stripped down version of #254, leaving out anything that isn’t strictly necessary. In particular, we do not touch the state tree, nor change the CBOR body of the HTTP requests and responses. In particular, the following (desireable) cleanup changes were left out:

 * It does _not_ introduce the `/canister_ranges` path in the state tree. This means a malicious edge node can swindle about the absence of requests.

   We can fix that later in a backward compatible way, so it is reasonable to leave it out.

 * `read_state` remains, and is _not_ replaced by a dedicated `call_status` endpoint.

   Since #254 was created, we have added other uses for `read_state` (#276), and also we want to get stateless routing without changes to more than the HTTP handler.

 * We do not move the location and encoding of call responses in the state tree. The motivation in #254 was to use this opportunity to reduce the amount of hashing that the replica has to do (by having fewer nodes in the call tree), make responding to these requests a bit, and maybe get better locality of state tree updates.

   We can still do that later in a transitional breaking change that doesn’t affect the edge nodes then.

 * In #254, we would have removed the odd `reply: { arg: bytes }` thing, as convenient drive-by cleanup to be done during a breaking change.

 * In #254, we would have wrapped the query response in `{ content : … }`, so that we can introduce query signatures in a backward compatible way. We may regret not doing that here, or we simpy introduce a new endpoint for “signed queries“ once we have them.

 * In #254, we would have changed `/api/v2/status` to allow for multiple spec versions to be listed in `ic_api_versions` (because we usually support multiple ones). Again, just a minor clean up that seemed opportune in a breaking change.

 * In #254, the `/call` endpoint should be invoked with `PUT`, because Nik says that’s what `PUT` is for.

(cherry picked from commit 047f4acf0c0349ff1c4041ac602969e9a650199a)